+++ plugins/frontend/backstage-plugin-jira/src/components/EntityJiraActivityStreamCard/index.ts
@@ -1,1 +1,2 @@
 export { EntityJiraActivityStreamCard } from './EntityJiraActivityStreamCard';
+export { ActivityStream } from './ActivityStream';
