+++ plugins/frontend/backstage-plugin-jira/src/components/JiraQueryCard/index.ts
@@ -1,0 +1,1 @@
+export * from './JiraQueryCard';
